Move functions in util.c that are used in the dissectors into a util-print.c.  The remaining functions are used only in the frontend.
